+++ pages/kickstarter/index.tsx
@@ -1,6 +1,6 @@
 import { NextPage } from "next"
 import RedirectPage from "~/components/RedirectPage"
 const Redirect: NextPage = () => (
-    <RedirectPage href="//www.kickstarter.com/projects/keesey/paleocene-4-comic-book" title="Kickstarter" />
+    <RedirectPage href="//www.kickstarter.com/projects/keesey/paleocene-5-comic-book" title="Kickstarter" />
 )
 export default Redirect