+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- release notes start -->
 
+## [2.11.1](https://github.com/keesey/keesey-comics.com/tree/v2.11.1) - 2025-06-02
+
+### Changed
+
+-   Updated Kickstarter link for _Paleocene #5_ campaign.
+-   _The Flesh #3_ is on sale, not coming soon.
+
 ## [2.11.0](https://github.com/keesey/keesey-comics.com/tree/v2.11.0) - 2025-04-07
 
 ### Changed